requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.31.2
anyio==4.12.1
billiard==4.2.4
bleach==6.3.0
blinker==1.9.0
blis==0.7.11
catalogue==2.0.10
celery==5.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
cryptography==42.0.8
cymem==2.0.13
Deprecated==1.3.1
distro==1.9.0
dnspython==2.6.1
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl#sha256=6a0f857a2b4d219c6fa17d455f82430b365bf53171a2d919b9376e5dc9be032e
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
filelock==3.25.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Limiter==4.1.1
flask-sock==0.7.0
flask-talisman==1.1.0
virustotal-python==1.1.1
fsspec==2026.2.0
google-ai-generativelanguage==0.6.10
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-generativeai==0.8.2
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.71.2
gunicorn==22.0.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
kombu==5.6.2
langcodes==3.5.1
limits==5.8.0
markdown-it-py==3.0.0
markdown_pdf==1.13.1
MarkupSafe==3.0.3
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
murmurhash==1.0.15
networkx==3.6.1
nltk==3.8.1
numpy==1.26.4
oauthlib==3.3.1
ordered-set==4.1.0
packaging==26.0
pandas==3.0.1
pillow==12.1.1
plantuml==0.3.0
preshed==3.0.12
prompt_toolkit==3.0.52
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.8.0
pymongo==4.7.3
PyMuPDF==1.27.1
PyMuPDFb==1.24.3
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
python-magic-bin==0.4.14
PyYAML==6.0.3
redis==5.0.7
regex==2026.2.28
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.5.1
scipy==1.17.1
sentence-transformers==3.0.1
setuptools==82.0.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smart_open==7.5.1
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
sympy==1.14.0
tbb==2021.13.1
thinc==8.2.5
threadpoolctl==3.6.0
tokenizers==0.19.1
torch==2.3.1
tqdm==4.67.3
transformers==4.42.4
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wasabi==1.1.3
wcwidth==0.6.0
weasel==0.4.3
webencodings==0.5.1
Werkzeug==3.0.3
wrapt==2.1.2
wsproto==1.3.2
